split config_linux to bpf and not bpf versions to prevent ebpf dependencies in non system-probe builds
